+++ src/pybamm/models/full_battery_models/sodium_ion/basic_dfn.py
@@ -181,9 +181,9 @@ class BasicDFN(pybamm.lithium_ion.BaseModel):
         ######################
         # Current in the solid
         ######################
-        sigma_eff_n = param.n.sigma(T) * eps_s_n**param.n.b_s
+        sigma_eff_n = param.n.sigma(T, sto_surf_n) * eps_s_n**param.n.b_s
         i_s_n = -sigma_eff_n * pybamm.grad(phi_s_n)
-        sigma_eff_p = param.p.sigma(T) * eps_s_p**param.p.b_s
+        sigma_eff_p = param.p.sigma(T, sto_surf_p) * eps_s_p**param.p.b_s
         i_s_p = -sigma_eff_p * pybamm.grad(phi_s_p)
         # The `algebraic` dictionary contains differential equations, with the key being
         # the main scalar variable of interest in the equation